+++ packages/react-core/src/components/Form/index.ts
@@ -2,3 +2,4 @@ export * from './ActionGroup';
 export * from './Form';
 export * from './FormGroup';
 export * from './FormHelperText';
+export * from './FormSection';