+++ packages/react-core/src/components/Nav/NavExpandable.tsx
@@ -6,9 +6,11 @@ import AngleRightIcon from '@patternfly/react-icons/dist/js/icons/angle-right-ic
 import { getUniqueId } from '../../helpers/util';
 import { NavContext } from './Nav';
 import { PickOptional } from '../../helpers/typeUtils';
+import { getOUIAProps, OUIAProps, getDefaultOUIAId } from '../../helpers';
 
 export interface NavExpandableProps
-  extends React.DetailedHTMLProps<React.LiHTMLAttributes<HTMLLIElement>, HTMLLIElement> {
+  extends React.DetailedHTMLProps<React.LiHTMLAttributes<HTMLLIElement>, HTMLLIElement>,
+    OUIAProps {
   /** Title shown for the expandable list */
   title: string;
   /** If defined, screen readers will read this text instead of the list title */
@@ -31,6 +33,7 @@ export interface NavExpandableProps
 
 interface NavExpandableState {
   expandedState: boolean;
+  ouiaStateId: string;
 }
 
 export class NavExpandable extends React.Component<NavExpandableProps, NavExpandableState> {
@@ -49,7 +52,8 @@ export class NavExpandable extends React.Component<NavExpandableProps, NavExpand
   id = this.props.id || getUniqueId();
 
   state = {
-    expandedState: this.props.isExpanded
+    expandedState: this.props.isExpanded,
+    ouiaStateId: getDefaultOUIAId(NavExpandable.displayName)
   };
 
   componentDidMount() {
@@ -90,10 +94,23 @@ export class NavExpandable extends React.Component<NavExpandableProps, NavExpand
   };
 
   render() {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { id, title, srText, children, className, isActive, groupId, isExpanded, onExpand, ...props } = this.props;
+    const {
+      title,
+      srText,
+      children,
+      className,
+      isActive,
+      ouiaId,
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      groupId,
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      id,
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      isExpanded,
+      ...props
+    } = this.props;
 
-    const { expandedState } = this.state;
+    const { expandedState, ouiaStateId } = this.state;
 
     const onClick = () => {
       this.setState(prevState => ({ expandedState: !prevState.expandedState }));
@@ -110,6 +127,7 @@ export class NavExpandable extends React.Component<NavExpandableProps, NavExpand
               isActive && styles.modifiers.current,
               className
             )}
+            {...getOUIAProps(NavExpandable.displayName, ouiaId !== undefined ? ouiaId : ouiaStateId)}
             onClick={(e: React.MouseEvent<HTMLLIElement, MouseEvent>) => this.handleToggle(e, context.onToggle)}
             {...props}
           >