+++ packages/react-core/src/components/TextArea/TextArea.tsx
@@ -84,3 +84,4 @@ export class TextAreaBase extends React.Component<TextAreaProps> {
 export const TextArea = React.forwardRef((props: TextAreaProps, ref: React.Ref<HTMLInputElement>) => (
   <TextAreaBase {...props} innerRef={ref as React.MutableRefObject<any>} />
 ));
+TextArea.displayName = 'TextArea';