+++ packages/react-core/src/components/TreeView/__tests__/TreeView.test.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import { TreeView } from '../TreeView';
-import { FolderIcon, FolderOpenIcon, FlagIcon } from '@patternfly/react-icons';
+import { Button } from '@patternfly/react-core';
+import { FolderIcon, FolderOpenIcon } from '@patternfly/react-icons';
 
 const options = [
   {
@@ -92,7 +93,11 @@ const flagOptions = [
     name: 'Cost Management',
     id: 'Cost',
     hasBadge: true,
-    action: <FolderIcon />,
+    action: (
+      <Button variant="plain" aria-label="Folder action">
+        <FolderIcon />
+      </Button>
+    ),
     children: [
       {
         name: 'Application 3',