+++ packages/react-integration/demo-app-ts/src/common/FakeRouterLink.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export const FakeRouterLink = ({ to, children, ...props }: any) => (
-  <a href={to} {...props}>
-    {children}
-  </a>
-);