+++ packages/react-integration/demo-app-ts/src/components/demos/LabelDemo/LabelDemo.tsx
@@ -1,7 +1,7 @@
 import { Label } from '@patternfly/react-core';
 import React, { Component } from 'react';
 import InfoCircleIcon from '@patternfly/react-icons/dist/js/icons/info-circle-icon';
-import { FakeRouterLink } from '../../../common/FakeRouterLink';
+import { Link } from 'react-router-dom';
 
 export class LabelDemo extends Component {
   componentDidMount() {
@@ -20,7 +20,9 @@ export class LabelDemo extends Component {
         <Label href="#" onClose={() => {}}>
           Grey link removeable
         </Label>
-        <Label variant="outline">Grey</Label>{' '}
+        <Label id="truncated-no-tooltip" variant="outline" isTruncated>
+          Grey
+        </Label>{' '}
         <Label variant="outline" icon={<InfoCircleIcon />}>
           Grey icon
         </Label>{' '}
@@ -40,21 +42,21 @@ export class LabelDemo extends Component {
         <Label color="blue" icon={<InfoCircleIcon />}>
           Blue icon
         </Label>
-        <Label color="blue" id="truncated-label" isTruncated icon={<InfoCircleIcon />}>
+        <Label color="blue" id="truncated-label" isTruncated tooltipPosition="top" icon={<InfoCircleIcon />}>
           Very very very very long label text that should be truncated
         </Label>
         <Label id="overflow-label" isOverflowLabel>
           Overflow label
         </Label>
         <Label
-          id="fake-router-link"
+          id="router-link"
           color="blue"
           icon={<InfoCircleIcon />}
           isTruncated
-          render={({ className, content }) => (
-            <FakeRouterLink to="/" className={className}>
+          render={({ className, content, componentRef }) => (
+            <Link to="/" className={className} innerRef={componentRef}>
               {content}
-            </FakeRouterLink>
+            </Link>
           )}
         >
           Blue label fake router link with icon that overflows