+++ packages/react-integration/demo-app-ts/src/components/demos/TopologyDemo/Connectors.tsx
@@ -1,7 +1,5 @@
 import * as React from 'react';
 import {
-  Visualization,
-  VisualizationSurface,
   Model,
   ModelKind,
   Edge,