+++ packages/react-datetime/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.15...@patternfly/react-datetime@4.2.16) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-datetime
+
+
+
+
+
+## [4.2.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.14...@patternfly/react-datetime@4.2.15) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-datetime
+
+
+
+
+
+## [4.2.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.13...@patternfly/react-datetime@4.2.14) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-datetime
+
+
+
+
+
+## [4.2.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.12...@patternfly/react-datetime@4.2.13) (2020-10-14)
+
+**Note:** Version bump only for package @patternfly/react-datetime
+
+
+
+
+
+## [4.2.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.11...@patternfly/react-datetime@4.2.12) (2020-10-13)
+
+**Note:** Version bump only for package @patternfly/react-datetime
+
+
+
+
+
+## [4.2.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.10...@patternfly/react-datetime@4.2.11) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-datetime
+
+
+
+
+
+## [4.2.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.9...@patternfly/react-datetime@4.2.10) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-datetime
+
+
+
+
+
+## [4.2.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.8...@patternfly/react-datetime@4.2.9) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-datetime
+
+
+
+
+
+## [4.2.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.7...@patternfly/react-datetime@4.2.8) (2020-10-09)
+
+**Note:** Version bump only for package @patternfly/react-datetime
+
+
+
+
+
+## [4.2.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.6...@patternfly/react-datetime@4.2.7) (2020-10-09)
+
+**Note:** Version bump only for package @patternfly/react-datetime
+
+
+
+
+
 ## [4.2.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-datetime@4.2.5...@patternfly/react-datetime@4.2.6) (2020-10-07)
 
 **Note:** Version bump only for package @patternfly/react-datetime