+++ packages/react-docs/CHANGELOG.md
@@ -3,6 +3,134 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.10.31](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.30...@patternfly/react-docs@5.10.31) (2020-10-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.30](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.29...@patternfly/react-docs@5.10.30) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.29](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.28...@patternfly/react-docs@5.10.29) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.28](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.27...@patternfly/react-docs@5.10.28) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.27](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.26...@patternfly/react-docs@5.10.27) (2020-10-14)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.26](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.25...@patternfly/react-docs@5.10.26) (2020-10-14)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.25](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.24...@patternfly/react-docs@5.10.25) (2020-10-13)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.24](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.23...@patternfly/react-docs@5.10.24) (2020-10-13)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.23](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.22...@patternfly/react-docs@5.10.23) (2020-10-13)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.22](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.21...@patternfly/react-docs@5.10.22) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.21](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.20...@patternfly/react-docs@5.10.21) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.19...@patternfly/react-docs@5.10.20) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.18...@patternfly/react-docs@5.10.19) (2020-10-12)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.17...@patternfly/react-docs@5.10.18) (2020-10-09)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.16...@patternfly/react-docs@5.10.17) (2020-10-09)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.15...@patternfly/react-docs@5.10.16) (2020-10-09)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [5.10.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.14...@patternfly/react-docs@5.10.15) (2020-10-08)
 
 **Note:** Version bump only for package @patternfly/react-docs